tests/wpt/web-platform-tests/fetch/api/headers/headers-errors.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Headers errors</title>
    <meta name="help" href="https://fetch.spec.whatwg.org/#headers">
    <meta name="author" title="Canon Research France" href="https://www.crf.canon.fr">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>

      test(function() {
        assert_throws(new TypeError() , function() { new Headers([["name"]]) });
      }, "Create headers giving an array having one string as init argument");

      test(function() {
        assert_throws(new TypeError() , function() { new Headers([["invalid", "invalidValue1", "invalidValue2"]]) });
      }, "Create headers giving an array having three strings as init argument");

      test(function() {
        assert_throws(new TypeError() , function() { new Headers([["invalidĀ", "Value1"]]) });
      }, "Create headers giving bad header name as init argument");

      test(function() {
        assert_throws(new TypeError() , function() { new Headers([["name", "invalidValueĀ"]]) });
      }, "Create headers giving bad header value as init argument");

      test(function() {
        var headers = new Headers();
        assert_throws(new TypeError() , function() { headers.get("invalidĀ") });
      }, "Check headers get with an invalid name");

      test(function() {
        var headers = new Headers();
        assert_throws(new TypeError() , function() { headers.delete("invalidĀ") });
      }, "Check headers delete with an invalid name");

      test(function() {
        var headers = new Headers();
        assert_throws(new TypeError() , function() { headers.has("invalidĀ") });
      }, "Check headers has with an invalid name");

      test(function() {
        var headers = new Headers();
        assert_throws(new TypeError() , function() { headers.set("invalidĀ", "Value1") });
      }, "Check headers set with an invalid name");

      test(function() {
        var headers = new Headers();
        assert_throws(new TypeError() , function() { headers.set("name", "invalidValueĀ") });
      }, "Check headers set with an invalid value");

      test(function() {
        var headers = new Headers();
        assert_throws(new TypeError() , function() { headers.append("invalidĀ", "Value1") });
      }, "Check headers append with an invalid name");

      test(function() {
        var headers = new Headers();
        assert_throws(new TypeError() , function() { headers.append("name", "invalidValueĀ") });
      }, "Check headers append with an invalid value");
    </script>
  </body>
</html>
